+++ GrAD/src/operations/unary_math/ADerf.hpp
@@ -41,7 +41,7 @@ vector<T> ADerf<T>::dfn(vector<T> x){
   vector<T> res(x.size());
   vector<T> dfnL = this->ptrL->dfn(x);
   T fnL = this->ptrL->fn(x);
-  for(int i = 0; i < x.size(); ++i){
+  for(int i = 0; (size_t)i < x.size(); ++i){
     res[i] = dfnL[i] * exp(-fnL*fnL) * T(2.0/sqrt(M_PI));
   }
   return res;
