+++ GrAD/src/operations/unary_math/ADlog2.hpp
@@ -40,7 +40,7 @@ vector<T> ADlog2<T>::dfn(vector<T> x){
   vector<T> res(x.size());
   vector<T> dfnL = this->ptrL->dfn(x);
   T fnL = this->ptrL->fn(x);
-  for(int i = 0; i < x.size(); ++i){
+  for(int i = 0; (size_t)i < x.size(); ++i){
     res[i] = dfnL[i] / (fnL * log(T(2.0)));
   }
   return res;
