+++ GrAD/src/operations/unary_math/ADsqrt.hpp
@@ -39,7 +39,7 @@ vector<T> ADsqrt<T>::dfn(vector<T> x){
   vector<T> res(x.size());
   vector<T> dfnL = this->ptrL->dfn(x);
   T fnL = this->ptrL->fn(x);
-  for(int i = 0; i < x.size(); ++i){
+  for(int i = 0; (size_t)i < x.size(); ++i){
     res[i] = - T(0.5) * dfnL[i] / sqrt(fnL);
   }
   return res;
